Writes metadata and title page to ebook
Converting first to markdown allows the script to add metadata to the
output format while prepending them to the output string.
This leads to a better formatted epub format, which can be read from a
Pocketbook reader without errors.

During this commit the first generated html text is better formatted now
and the title is removed from the html text, because it is added to the
resulting document via metadata and written on the title page.